refactorise builder and schema for better structure
