+++ GraphQL_FS/graphql-server/src/models/User.ts
@@ -1,4 +1,4 @@
-import { builder } from '../schema';
+import { builder } from '../builder';
 import { prisma } from '../db';
 
 builder.prismaObject('User', {
