+++ GraphQL_FS/graphql-server/src/schema.ts
@@ -1,26 +1,6 @@
-import SchemaBuilder from '@pothos/core';
-import { DateResolver } from 'graphql-scalars';
-import PrismaPlugin from '@pothos/plugin-prisma';
-import type PrismaTypes from '@pothos/plugin-prisma/generated';
-import { prisma } from './db';
+import { builder } from './builder';
 
-//import Scalars Date type to be used in the Schemabuilder
-export const builder = new SchemaBuilder<{
-  Scalars: {
-    Date: { Input: Date; Output: Date };
-  };
-  PrismaTypes: PrismaTypes;
-}>({
-  plugins: [PrismaPlugin],
-  prisma: {
-    client: prisma,
-  },
-});
-//builder require a prisma instance
+import './models/Message';
+import './models/User';
 
-builder.addScalarType('Date', DateResolver, {});
-
-//initialize the query type
-builder.queryType({});
-
-export const schema = builder.toSchema();
+export const schema = builder.toSchema({});
